pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Ceferino Patino", email = "c4patino@gmail.com" }
]
license = {text = "MIT"}
readme = "README.md"

requires-python = "~3.12"
dependencies = [
    "django (==5.0.2)",
    "django-cors-headers (==4.3.1)",
    "django-cleanup (==8.1.0)",
    "django-q2 (==1.6.2)",
    "django-rest-framework (==0.1.0)",
    "django-prometheus (==2.3.1)",
    "django-sass-processor (==1.4)",
    "drf-nested-routers (==0.93.5)",
    "drf-spectacular (==0.27.1)",
    "drf-spectacular-sidecar (==2024.2.1)",
    "lightning (==2.2.0)",
    "matplotlib (==3.8.3)",
    "mysqlclient (==2.2.4)",
    "numpy (==1.26.4)",
    "opencv-python (==4.9.0.80)",
    "pandas (==2.2.1)",
    "psutil (==5.9.8)",
    "pyarrow (==15.0.0)",
    "rich (==13.7.1)",
    "redis (==5.0.3)",
    "scikit-image (==0.22.0)",
    "tensorboard (==2.16.2)",
    "torchvision (==0.17.1)",
    "whitenoise (==6.6.0)",
    "ultralytics (>=8.3.78,<9.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-complexity = 25

[tool.yapf]
COLUMN_LIMIT = 130
